Refactor optimization and single-point energy calculation: - Update `run_opt` to use `Path` for file handling and improve readability. - Move `read_xyz_file` function to a new utility module for better organization.
