+++ src/pymars/phoboos/singlepoint.py
@@ -1,5 +1,4 @@
 import numpy as np
-import os
 import jax.numpy as jnp
 from pathlib import Path
 
@@ -7,42 +6,7 @@ from fennol.models import FENNIX
 from fennol.utils.periodic_table import PERIODIC_TABLE_REV_IDX
 from pymars.utils import format_batch_conformations, us
 
-def read_xyz_file(xyz_path: str) -> tuple[list[str], np.ndarray]:
-    """Read a single-frame XYZ file."""
-    if not os.path.exists(xyz_path):
-        raise FileNotFoundError(f"XYZ file not found: {xyz_path}")
-
-    species = []
-    coords = []
-
-    with open(xyz_path, "r") as f:
-        lines = f.readlines()
-
-    if len(lines) < 2:
-        raise ValueError(f"XYZ file {xyz_path} has fewer than 2 lines")
-
-    try:
-        n_atoms = int(lines[0].strip())
-        #print(f"Number of atoms: {n_atoms}")
-    except ValueError:
-        raise ValueError(f"First line must be an integer (number of atoms)")
-
-    if len(lines) < 2 + n_atoms:
-        raise ValueError(f"XYZ file claims {n_atoms} atoms but has only {len(lines) - 2} data lines")
-
-    for i in range(2, 2 + n_atoms):
-        parts = lines[i].split()
-        if len(parts) < 4:
-            raise ValueError(f"Line {i+1} has fewer than 4 fields")
-        sym = parts[0]
-        try:
-            x, y, z = float(parts[1]), float(parts[2]), float(parts[3])
-        except ValueError:
-            raise ValueError(f"Line {i+1} has non-numeric coordinates")
-        species.append(sym)
-        coords.append([x, y, z])
-
-    return species, np.array(coords, dtype=np.float64)
+from pymars.phoboos.utils import read_xyz_file
 
 def run_spt(xyz_file, model_file, outfile=None, total_charge=0):
     "Calculate the energy of a molecular geometry using a FENNIX model"
